fixed some VS14 compiler warnings

+++ src/sbml/packages/multi/extension/test/TestRunner.c
@@ -47,7 +47,7 @@ void
 setTestDataDirectory (void)
 {
   char *srcdir = getenv("srcdir");
-  int  length  = (srcdir == NULL) ? 0 : strlen(srcdir);
+  int  length  = (srcdir == NULL) ? 0 : static_cast<int>( strlen(srcdir) );
 
 
   /**
